ms-notification/src/main/java/fr/shoppo/msnotification/domain/bo/commerce/Commercant.java
package fr.shoppo.msnotification.domain.bo.commerce;

import fr.shoppo.msnotification.domain.bo.Template;

public class Commercant implements Template {
    Integer id;
    String email;
    String password;
    Commerce commerce;
    String name;
    Boolean owner;

    public Commercant() {
        /*data binding*/
    }

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public Commerce getCommerce() {
        return commerce;
    }

    public void setCommerce(Commerce commerce) {
        this.commerce = commerce;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Boolean getOwner() {
        return owner;
    }

    public void setOwner(Boolean owner) {
        this.owner = owner;
    }

    @Override
    public String toString() {
        return "{" +
                "id=" + id +
                ", email='" + email + '\'' +
                ", password='" + password + '\'' +
                ", commerce=" + commerce +
                '}';
    }
}
